Corrected plots and Added cells for testing errorbars
Corrected plots in NewVersion4_Corrected.ipynb and added a few cells at
the end of NewVersion4.ipynb for testing errorbars in final graphs
(with weights using scipy.odr). The script Delete.ipynb was created to
test example of the use of scipy.odr (delete when not needed anymore).
The notebook NewVersion4_Testing.ipynb should probably be deleted.